Object-Mode/FullMix-Mode
Allows to choose between mono-channel (object mode) and full-channel
(bed mode) to be send to JACK. Set the soundcard speaker setup to
the highest channel layout you want to use for the bed, set Unity's
speaker layout to that same bed layout and then decide for each
channel in the mixer the mode.

Set the JACK channel indexes appropriately!